revert: use binary_search_by instead of find + clean: move completion tests to source_info
